+++ raiden-ts/src/transfers/epics/withdraw.ts
@@ -32,7 +32,8 @@ import { getCap } from '../../transport/utils';
 import { chooseOnchainAccount, getContractWithSigner } from '../../helpers';
 import { withdrawExpire, withdrawMessage, withdraw, withdrawCompleted } from '../actions';
 import { Direction } from '../state';
-import { retrySendUntil$, matchWithdraw, exponentialBackoff } from './utils';
+import { intervalFromConfig } from '../../config';
+import { retrySendUntil$, matchWithdraw } from './utils';
 
 /**
  * Emits withdraw action once for each own non-confirmed message at startup
@@ -101,8 +102,7 @@ export function withdrawSendRequestMessageEpic(
   return action$.pipe(
     filter(withdrawMessage.request.is),
     filter((action) => action.meta.direction === Direction.SENT),
-    withLatestFrom(config$),
-    mergeMap(([action, { pollingInterval, httpTimeout }]) => {
+    mergeMap((action) => {
       const message = action.payload.message;
       const send = messageSend.request(
         { message },
@@ -125,12 +125,7 @@ export function withdrawSendRequestMessageEpic(
         }),
       );
       // emit request once immediatelly, then wait until success, then retry every 30s
-      return retrySendUntil$(
-        send,
-        action$,
-        notifier,
-        exponentialBackoff(pollingInterval, httpTimeout * 2),
-      );
+      return retrySendUntil$(send, action$, notifier, intervalFromConfig(config$));
     }),
   );
 }
@@ -148,12 +143,22 @@ export function withdrawSendRequestMessageEpic(
  * @param deps.log - Logger instance
  * @param deps.getTokenNetworkContract - TokenNetwork contract getter
  * @param deps.latest$ - Latest observable
+ * @param deps.config$ - Config observable
  * @returns Observable of withdrawExpired actions
  */
 export function withdrawSendTxEpic(
   action$: Observable<RaidenAction>,
   {}: Observable<RaidenState>,
-  { address, signer, main, provider, log, getTokenNetworkContract, latest$ }: RaidenEpicDeps,
+  {
+    address,
+    signer,
+    main,
+    provider,
+    log,
+    getTokenNetworkContract,
+    latest$,
+    config$,
+  }: RaidenEpicDeps,
 ): Observable<withdraw.success | withdraw.failure> {
   return action$.pipe(
     filter(withdrawMessage.success.is),
@@ -200,7 +205,7 @@ export function withdrawSendTxEpic(
               log,
               provider,
             }),
-            retryWhile(provider.pollingInterval, { onErrors: commonTxErrors, log: log.debug }),
+            retryWhile(intervalFromConfig(config$), { onErrors: commonTxErrors, log: log.debug }),
           );
         }),
         map(([, receipt]) =>
@@ -348,8 +353,7 @@ export function withdrawSendExpireMessageEpic(
   return action$.pipe(
     filter(withdrawExpire.success.is),
     filter((action) => action.meta.direction === Direction.SENT),
-    withLatestFrom(config$),
-    mergeMap(([action, { pollingInterval, httpTimeout }]) => {
+    mergeMap((action) => {
       const message = action.payload.message;
       const send = messageSend.request(
         { message },
@@ -369,12 +373,7 @@ export function withdrawSendExpireMessageEpic(
       );
       // besides using notifier to stop retry, also merge the withdrawCompleted output action
       return merge(
-        retrySendUntil$(
-          send,
-          action$,
-          notifier,
-          exponentialBackoff(pollingInterval, httpTimeout * 2),
-        ),
+        retrySendUntil$(send, action$, notifier, intervalFromConfig(config$)),
         notifier,
       );
     }),