+++ raiden-ts/src/utils/ethers.ts
@@ -8,20 +8,44 @@ import type {
   Log,
   Network,
 } from '@ethersproject/providers';
-import { Observable, fromEventPattern, from, EMPTY, defer } from 'rxjs';
-import { mergeMap, debounceTime, catchError, exhaustMap } from 'rxjs/operators';
+import { Observable, fromEventPattern, from, defer } from 'rxjs';
+import { mergeMap, debounceTime, exhaustMap, concatMap, tap } from 'rxjs/operators';
 
-import { isntNil } from './types';
+import { networkErrors } from './error';
+import { retryWhile } from './rx';
+
+/**
+ * @param provider - Provider to getLogs from
+ * @param filter - getLogs filter
+ * @param chunk - Chunk size
+ * @returns Observable of fetched logs
+ */
+export function getLogsByChunk$(
+  provider: JsonRpcProvider,
+  filter: Filter & { fromBlock: number; toBlock: number },
+  chunk = 1e5,
+): Observable<Log> {
+  const { fromBlock, toBlock } = filter;
+  const ranges: [number, number][] = [];
+  for (let i = fromBlock; i <= toBlock; i += chunk)
+    ranges.push([i, Math.min(i + chunk - 1, toBlock)]);
+  return from(ranges).pipe(
+    concatMap(([fromBlock, toBlock]) =>
+      defer(async () => provider.getLogs({ ...filter, fromBlock, toBlock })).pipe(
+        retryWhile(provider.pollingInterval, { onErrors: networkErrors }),
+        mergeMap((logs) => from(logs)),
+      ),
+    ),
+  );
+}
 
 export function fromEthersEvent<T>(
   target: JsonRpcProvider,
   event: string | string[],
-  resultSelector?: (...args: any[]) => T,
 ): Observable<T>;
 export function fromEthersEvent<T extends Log>(
   target: JsonRpcProvider,
   event: Filter,
-  resultSelector?: (...args: any[]) => T,
   confirmations?: number,
   fromBlock?: number,
 ): Observable<T>;
@@ -30,8 +54,6 @@ export function fromEthersEvent<T extends Log>(
  *
  * @param target - Object to hook event listener, maybe a Provider or Contract
  * @param event - EventFilter or string representing the event to listen to
- * @param resultSelector - A map of events arguments to output parameters
- *      Default is to pass only first parameter
  * @param confirmations - After how many blocks a tx is considered confirmed
  * @param fromBlock - Block since when to fetch events from
  * @returns Observable of target.on(event) events
@@ -39,7 +61,6 @@ export function fromEthersEvent<T extends Log>(
 export function fromEthersEvent<T>(
   target: JsonRpcProvider,
   event: EventType,
-  resultSelector?: (...args: any[]) => T,
   confirmations = 5,
   fromBlock?: number,
 ) {
@@ -47,7 +68,6 @@ export function fromEthersEvent<T>(
     return fromEventPattern<T>(
       (handler: Listener) => target.on(event, handler),
       (handler: Listener) => target.removeListener(event, handler),
-      resultSelector,
     ) as Observable<T>;
 
   const range = confirmations * 2; // half for confirmed, half for unconfirmed logs
@@ -65,7 +85,8 @@ export function fromEthersEvent<T>(
           : 1;
       fromBlock = resetBlock - confirmations;
     }
-    blockQueue.push(fromBlock); // starts 'blockQueue' with subscription-time's resetEventsBlock
+    // starts 'blockQueue' with subscription-time's resetEventsBlock
+    blockQueue.splice(0, blockQueue.length, fromBlock);
 
     return fromEthersEvent<number>(target, 'block');
   }).pipe(
@@ -73,36 +94,26 @@ export function fromEthersEvent<T>(
     // exhaustMap will skip new events if it's still busy with a previous getLogs call,
     // but next [fromBlock] in queue always includes range for any skipped block
     exhaustMap((blockNumber) =>
-      defer(() =>
-        target.getLogs({ ...event, fromBlock: blockQueue[0], toBlock: blockNumber }),
-      ).pipe(
-        mergeMap((logs) => {
-          // if queue is full, pop_front 'fromBlock' which was just queried
-          if (blockQueue.length >= range) blockQueue.shift();
-          // push_back next block iff getLogs didn't throw, queue is never empty
-          blockQueue.push(blockNumber + 1);
-
-          // if a confirmed log comes, clear queued smaller blocks and push_front block after
-          // let unconfirmed actions be emitted multiple times to update pendingTxs if needed
-          const afterLogBlock =
-            Math.max(
-              0,
-              ...logs
-                .map((log) => log.blockNumber)
-                .filter(isntNil)
-                .filter((block) => block + confirmations <= blockNumber), // only confirmed
-            ) + 1;
-          if (blockQueue[0] <= afterLogBlock)
-            blockQueue.splice(
-              0, // from queue's front
-              blockQueue.filter((block) => block <= afterLogBlock).length, // clear older blocks
-              afterLogBlock, // push_front block after newest log's blockNumber
-            );
-
-          return from(logs); // unwind logs
+      getLogsByChunk$(target, { ...event, fromBlock: blockQueue[0], toBlock: blockNumber }).pipe(
+        tap({
+          next: (log) => {
+            // don't clear blockQueue for non-confirmed logs
+            if (!log.blockNumber || log.blockNumber + confirmations > blockNumber) return;
+            const nextBlock = log.blockNumber + 1;
+            // index of first block which should stay on the queue;
+            let clearHead = blockQueue.findIndex((b) => b > nextBlock);
+            if (!clearHead) return;
+            else if (clearHead < 0) clearHead = blockQueue.length; // clear whole queue
+            blockQueue.splice(0, clearHead, nextBlock);
+            // invariant: blockQueue length never increases here
+          },
+          complete: () => {
+            // if queue is full, pop_front 'fromBlock' which was just queried
+            while (blockQueue.length >= range) blockQueue.shift();
+            // push_back next block iff getLogs didn't throw, queue is never empty
+            blockQueue.push(blockNumber + 1);
+          },
         }),
-        // `getLogs` errors skip [blockQueue] update, so previous fromBlock is retried later
-        catchError(() => EMPTY),
       ),
     ),
   );