+++ raiden-ts/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 ### Added
 - [#211] 'suggestPartners' method to fetch suggested partners from PFS
+- [#2417] Make 'start' async, introduce 'synced' promise, both resolves when syncing finishes
 
 [#211]: https://github.com/raiden-network/light-client/issues/211
+[#2417]: https://github.com/raiden-network/light-client/pull/2417
 
 ## [0.14.0] - 2020-11-25
 ### Fixed